app_repo/requirements.txt
flask
requests
paramiko
kafka-python

